+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.1
+
+* Added tables support.
+* Made a code block scrollable.
+* Changed images span default alignment.
+* Fix whitespaces trimming in markdown texts.
+
 ## 0.1.0
 
 * Added support for Web.